RMQLookupTable: reduced memory for smaller inputs

+++ src/main/java/com/ugav/algo/RMQPlusMinusOne.java
@@ -60,7 +60,7 @@ public class RMQPlusMinusOne extends RMQLinearAbstract {
 	for (int b = 0; b < ds.blockNum; b++) {
 	    int key = calcBlockKey(ds, b);
 
-	    ds.interBlocksDs[b] = tables.computeIfAbsent(Integer.valueOf(key), k -> {
+	    ds.interBlocksDs[b] = tables.computeIfAbsent(key, k -> {
 		int demoBlock[] = calcDemoBlock(k, ds.blockSize);
 		return RMQLookupTable.getInstace().preprocessRMQ(new IntArrayComperator(demoBlock), demoBlock.length);
 	    });
@@ -78,7 +78,7 @@ public class RMQPlusMinusOne extends RMQLinearAbstract {
 
 	@Override
 	int getBlockSize(int n) {
-	    return (int) Math.ceil(Utils.log2(n) / 2);
+	    return (int) Math.ceil(Utils.log2((double) n) / 2);
 	}
 
 	@Override
